requirements.txt
dj-database-url==0.5.0
Django==2.1.4
future==0.17.1
idna==2.7
ipython==7.2.0
oauthlib==2.1.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-decouple==3.1
python-twitter==3.5
pytz==2018.7
requests==2.20.1